requirements.txt
colour
ipython>=8.18.0
isosurfaces
fontTools
manimpango>=0.6.0
mapbox-earcut
matplotlib
moderngl
moderngl_window
numpy
Pillow
pydub
pygments
PyOpenGL
pyperclip
pyrr
pyyaml
rich
scipy
screeninfo
skia-pathops
svgelements>=1.8.1
sympy
tqdm
typing-extensions; python_version < "3.11"
validators
Sphinx==3.0.3
sphinx-copybutton
furo==2020.10.5b9
Jinja2